config autocompletar usuario agendamento e probir deletar caso haja liberação de veículo cinculada

+++ Agenda/views.py
@@ -1,9 +1,11 @@
+from urllib import request
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.contrib.auth.decorators import login_required
 from django.views import generic
 from django.views.generic import ListView
 from django.urls import reverse_lazy
 from django.shortcuts import redirect
+from django.http import HttpResponseRedirect
 from Agenda.models import Agenda
 from Agenda.filters import AgendaFilter
 from Agenda import forms
@@ -15,7 +17,7 @@ from sweetify.views import SweetifySuccessMixin
 class AgendaRegisterView(SweetifySuccessMixin, generic.CreateView, LoginRequiredMixin):
     model = Agenda
     fields = ['descricao', 'data_saida', 'data_retorno', 'hora_saida',
-              'hora_retorno', 'veiculo', 'motorista', 'usuario_cadastro']
+              'hora_retorno', 'veiculo', 'motorista']
     success_message = 'Cadastrado!'
     sweetify_options = {'text': 'Informações do Agendamento cadastradas com sucesso.',
                         'timer': 2500
@@ -27,6 +29,17 @@ class AgendaRegisterView(SweetifySuccessMixin, generic.CreateView, LoginRequired
         context = super(AgendaRegisterView, self).get_context_data(**kwargs)
         context['cadastrar_agendamento'] = 'active'
         return context
+
+    def form_valid(self, form):
+        self.object = form.save(commit=False)
+        self.object.usuario_cadastro = self.request.user
+        self.object.save()
+        response = super(SweetifySuccessMixin, self).form_valid(form)
+        success_message = self.get_success_message(form.cleaned_data)
+        if success_message:
+            sweetify.success(self.request, success_message,
+                             **self.get_sweetify_options())
+        return response
 
 
 class AgendaListView(SweetifySuccessMixin, ListView, LoginRequiredMixin):
@@ -54,7 +67,7 @@ class AgendaEditView(SweetifySuccessMixin, generic.UpdateView, LoginRequiredMixi
     template_name = 'Agenda/cadastrar_agendamento.html'
     success_message = 'Alterado com Sucesso!'
     sweetify_options = {'text': 'Informações do Agendamento alteradas com sucesso.',
-                        'timer': 2500
+                        'timer': 2000
                         }
 
     def get_success_url(self):
@@ -63,9 +76,13 @@ class AgendaEditView(SweetifySuccessMixin, generic.UpdateView, LoginRequiredMixi
 
 @login_required
 def agendamento_delete(request, pk):
-    if id:
+    agenda = Agenda.objects.get(id=pk)
+    if id and not agenda.agendamento_liberacao_veiculo.all():
         Agenda.objects.filter(id=pk).delete()
         sweetify.success(request, 'Cadastro Excluído ',
                          text='Informações do Agendamento excluídas com \
                                      sucesso.', timer=2000)
+    else:
+        sweetify.error(request, 'Erro ao excluir ',
+                       text='Este Agendamento já possui uma Liberação cadastrada', timer=3000)
     return redirect("agenda:listar_agendamentos")
